--- src/main/java/kr/bb/store/domain/common/entity/Sido.java
@@ -1,18 +1,0 @@
-package kr.bb.store.domain.common.entity;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Entity
-public class Sido {
-    @Id
-    private String code;
-
-    private String name;
-}
