+++ src/main/java/kr/bb/store/domain/store/entity/address/Gugun.java
@@ -1,0 +1,21 @@
+package kr.bb.store.domain.store.entity.address;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Entity
+public class Gugun {
+    @Id
+    private String code;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "sido_code", nullable = false)
+    private Sido sido;
+
+    private String name;
+}
