+++ src/main/java/kr/bb/store/domain/store/entity/address/Sido.java
@@ -1,21 +1,18 @@
-package kr.bb.store.domain.common.entity;
+package kr.bb.store.domain.store.entity.address;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.Id;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 @Entity
-public class Gugun {
+public class Sido {
     @Id
     private String code;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "sido_code", nullable = false)
-    private Sido sido;
 
     private String name;
 }
